added first water surface attempt... (no waves, no refraction, no fresnel..., just a flat mirror...)

+++ trunk/PGR2project/src/Light.cpp
@@ -16,7 +16,7 @@ void Light::init()
 	
 }
 
-void Light::update(float time)
+void Light::update(double time)
 {
 
 }
